子查询替换、2

+++ src/main/java/com/exercise/mybatis/dao/Dir.java
@@ -13,6 +13,6 @@ public class Dir {
     private Long id;
     private Long parentId;
     private String dirName;
-    private Long isAddDir;
+    private Integer isAddDir;
     private List<Dir> children;
 }
